VictoryScreen reorg - make helpers reusable
