+++ core/src/com/unciv/ui/components/TabbedPager.kt
@@ -9,6 +9,7 @@ import com.badlogic.gdx.scenes.scene2d.InputListener
 import com.badlogic.gdx.scenes.scene2d.ui.Button
 import com.badlogic.gdx.scenes.scene2d.ui.Cell
 import com.badlogic.gdx.scenes.scene2d.ui.Image
+import com.badlogic.gdx.scenes.scene2d.ui.Label
 import com.badlogic.gdx.scenes.scene2d.ui.ScrollPane
 import com.badlogic.gdx.scenes.scene2d.ui.Table
 import com.badlogic.gdx.scenes.scene2d.ui.WidgetGroup
@@ -24,6 +25,7 @@ import com.unciv.ui.components.extensions.keyShortcuts
 import com.unciv.ui.components.extensions.onActivation
 import com.unciv.ui.components.extensions.packIfNeeded
 import com.unciv.ui.components.extensions.pad
+import com.unciv.ui.components.extensions.toLabel
 import com.unciv.ui.images.IconTextButton
 import com.unciv.ui.popups.Popup
 import com.unciv.ui.screens.basescreen.BaseScreen
@@ -105,6 +107,40 @@ open class TabbedPager(
 
         /** @return Optional second content [Actor], will be placed outside the tab's main [ScrollPane] between header and `content`. Scrolls horizontally only. */
         fun getFixedContent(): Actor? = null
+
+        /** Sets first row cell's minWidth to the max of the widths of that column over all given tables
+         *
+         * Notes:
+         * - This aligns columns only if the tables are arranged vertically with equal X coordinates.
+         * - first table determines columns processed, all others must have at least the same column count.
+         * - Tables are left as needsLayout==true, so while equal width is ensured, you may have to pack if you want to see the value before this is rendered.
+         */
+        fun equalizeColumns(vararg tables: Table) {
+            for (table in tables)
+                table.packIfNeeded()
+            val columns = tables.first().columns
+            if (tables.any { it.columns < columns })
+                throw IllegalStateException("IPageExtensions.equalizeColumns needs all tables to have at least the same number of columns as the first one")
+            val widths = (0 until columns)
+                .mapTo(ArrayList(columns)) { column ->
+                    tables.maxOf { it.getColumnWidth(column) }
+                }
+            for (table in tables) {
+                for (column in 0 until columns)
+                    table.cells[column].run {
+                        if (actor == null)
+                        // Empty cells ignore minWidth, so just doing Table.add() for an empty cell in the top row will break this. Fix!
+                            setActor<Label>("".toLabel())
+                        else if (Align.isCenterHorizontal(align)) (actor as? Label)?.run {
+                            // minWidth acts like fillX, so Labels will fill and then left-align by default. Fix!
+                            if (!Align.isCenterHorizontal(labelAlign))
+                                setAlignment(Align.center)
+                        }
+                        minWidth(widths[column] - padLeft - padRight)
+                    }
+                table.invalidate()
+            }
+        }
     }
 
     //endregion
